# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "ShoppingCartResponse": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "ownerId": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/Array"
                        },
                        "type": "array"
                    },
                    "updatedAt": {
                        "type": "number"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "ownerId",
                    "total",
                    "items",
                    "updatedAt"
                ]
            },
...
            "SellerItemCreationRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 75,
                        "minLength": 5
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 25
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0
                    },
                    "categoryId": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "sellerId": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    },
                    "quantity": {
                        "type": "number",
                        "minimum": 0
                    },
                    "imageUrl": {
                        "format": "url",
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "price",
                    "categoryId",
                    "sellerId"
                ]
            },
            "SellerItemUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 75,
                        "minLength": 5
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 25
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0
                    },
                    "categoryId": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    },
                    "quantity": {
                        "type": "number",
                        "minimum": 0
                    },
                    "imageUrl": {
                        "format": "url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CartItemRequest": {
                "properties": {
                    "id": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number",
                        "minimum": 1
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "quantity"
                ]
            },
...
        }
    },
    "info": {
        "title": "sustainable-community-server API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/cart/": {
            "get": {
                "operationId": "ShoppingCartController.getCart",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShoppingCartResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get shopping cart",
                "tags": [
                    "Shopping Cart"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "ShoppingCartController.clearCart",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Clear shopping cart",
                "tags": [
                    "Shopping Cart"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items": {
            "post": {
                "operationId": "ShoppingCartController.addItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItemRequest"
                            }
                        }
                    },
                    "description": "CartItemRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Add item to shopping cart",
                "tags": [
                    "Shopping Cart"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "operationId": "ShoppingCartController.updateItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItemRequest"
                            }
                        }
                    },
                    "description": "CartItemRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update item in shopping cart",
                "tags": [
                    "Shopping Cart"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items/{itemId}": {
            "delete": {
                "operationId": "ShoppingCartController.removeItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "itemId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Remove item from shopping cart",
                "tags": [
                    "Shopping Cart"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/items/": {
            "get": {
                "operationId": "SellerItemController.getListOfItems",
                "parameters": [
                    {
                        "in": "query",
                        "name": "sellerId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "categoryId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "isAvailable",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SellerItemResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get list of items",
                "tags": [
                    "Seller Item"
                ]
            },
            "post": {
                "operationId": "SellerItemController.createItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SellerItemCreationRequest"
                            }
                        }
                    },
                    "description": "SellerItemCreationRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerItemResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Create an item",
                "tags": [
                    "Seller Item"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/{id}": {
            "get": {
                "operationId": "SellerItemController.getItem",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerItemResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get item by id",
                "tags": [
                    "Seller Item"
                ]
            },
            "put": {
                "operationId": "SellerItemController.updateItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SellerItemUpdateRequest"
                            }
                        }
                    },
                    "description": "SellerItemUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerItemResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Update an item",
                "tags": [
                    "Seller Item"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "SellerItemController.deleteItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete an item",
                "tags": [
                    "Seller Item"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}